kurbo

Module bezpath

Bézier paths (up to cubic).

Structs

BezPath

A Bézier path.

CloseSubpaths 🔒

An iterator that closes all open subpaths.

LineIntersection

An intersection of a Line and a PathSeg.

MinDistance

The minimum distance between two Bézier curves.

PathSegIter

An iterator for path segments.

Segments

An iterator that transforms path elements to path segments.

Enums

CloseSubpathState 🔒

PathEl

The element of a Bézier path.

PathSeg

A segment of a Bézier path.

Constants

TO_QUAD_TOL 🔒

Proportion of tolerance budget that goes to cubic to quadratic conversion.

Functions

close_subpaths 🔒

Close all open subpaths in a path, expressed as iterator transform.

cubic_bez_coefs 🔒

flatten

Flatten the path, invoking the callback repeatedly.

quadratic_bez_coefs 🔒

reverse_subpath 🔒

Helper for reversing a subpath.

segments

Transform an iterator over path elements into one over path segments.