kurbo

Module ellipse

Implementation of ellipse shape.

Structs

Ellipse

An ellipse.

Functions

agm_elliptic_perimeter 🔒

Calculates circumference C of an ellipse with radii (x, y) using the arithmetic-geometric mean, as described in equation 19.8.6 of https://web.archive.org/web/20240926233336/https://dlmf.nist.gov/19.8#i.

kummer_elliptic_perimeter 🔒

Calculates circumference C of an ellipse with radii (x, y) as the infinite series

kummer_elliptic_perimeter_range 🔒

This calculates the error range of kummer_elliptic_perimeter. That function returns a lower bound on the true value, and though we do not know what the remainder of the infinite series sums to, we can calculate an upper bound: