Enum kurbo::PathEl

pub enum PathEl {
    MoveTo(Point),
    LineTo(Point),
    QuadTo(Point, Point),
    CurveTo(Point, Point, Point),
    ClosePath,
}

The element of a Bรฉzier path.

A valid path has MoveTo at the beginning of each subpath.

Variants

MoveTo(Point)

Move directly to the point without drawing anything, starting a new subpath.

LineTo(Point)

Draw a line from the current location to the point.

QuadTo(Point, Point)

Draw a quadratic bezier using the current location and the two points.

CurveTo(Point, Point, Point)

Draw a cubic bezier using the current location and the three points.

ClosePath

Close off the path.

Implementations

impl PathEl

pub fn is_finite(&self) -> bool

Is this path element finite?

pub fn is_nan(&self) -> bool

Is this path element NaN?