Enum kurbo::PathSeg

pub enum PathSeg {
    Line(Line),
    Quad(QuadBez),
    Cubic(CubicBez),
}

A segment of a Bézier path.

Variants

Line(Line)

Quad(QuadBez)

Cubic(CubicBez)

Methods

impl PathSeg

pub fn reverse(&self) -> PathSeg

Returns a new PathSeg describing the same path as self, but with the points reversed.

pub fn to_cubic(&self) -> CubicBez

Convert this segment to a cubic bezier.

fn winding_inner(&self, p: Point) -> i32

fn winding(&self, p: Point) -> i32

Compute the winding number contribution of a single segment.

Cast a ray to the left and count intersections.