kurbo::expand

Struct ExpandCtx

struct ExpandCtx {
    expand: Diagonal2,
    join: Join,
    miter_limit: f64,
    tolerance: f64,
    result: BezPath,
    first_n: Option<Vec2>,
    first_tan: Vec2,
    last_pt: Point,
    last_n: Option<Vec2>,
    last_tan: Vec2,
}

Fields
expand: Diagonal2join: Joinmiter_limit: f64tolerance: f64result: BezPathfirst_n: Option<Vec2>first_tan: Vec2last_pt: Pointlast_n: Option<Vec2>last_tan: Vec2
Implementations

impl ExpandCtx

fn in_tolerance(&self, v: Vec2) -> bool

Helper function to determine if a distance is within tolerance.

fn do_line(&mut self, p1: Point)

Process a line segment. Includes initial join.

fn do_quad(&mut self, p1: Point, p2: Point)

Process a quadratic Bézier segment. Includes initial join.

fn do_cubic(&mut self, p1: Point, p2: Point, p3: Point)

Process a cubic segment. Includes initial join.

fn do_join(&mut self, n: Vec2, tan: Vec2, is_line: bool)

Do a join.

The tan parameter is a vector tangent to the start of the new segment. The n parameter is the normal vector (turned tangent) scaled by the expansion.

fn do_close_path(&mut self, first_pt: Point)

Close an open subpath if there is one; no-op if already closed.