kurbo

Function stroke_with

pub fn stroke_with(
    path: impl IntoIterator<Item = PathEl>,
    style: &Stroke,
    _opts: &StrokeOpts,
    tolerance: f64,
    ctx: &mut StrokeCtx,
)

Expand a stroke into a fill.

This is the same as stroke, except for the fact that you can explicitly pass a StrokeCtx. By doing so, you can reuse the same context over multiple calls and ensure that the number of reallocations is minimized.

Unlike stroke, this method doesn’t return an owned version of the expanded stroke as a BezPath. Instead, you can get a reference to the resulting path by calling StrokeCtx::output.