Module kurbo::param_curve

A trait for curves parametrized by a scalar.

Constants

MAX_EXTREMA
The maximum number of extrema that can be reported in the ParamCurveExtrema trait.

Traits

ParamCurve
A curve parametrized by a scalar.

ParamCurveArclen
A parametrized curve that can have its arc length measured.

ParamCurveArea
A parametrized curve that can have its signed area measured.

ParamCurveCurvature
A parametrized curve that reports its curvature.

ParamCurveDeriv
A differentiable parametrized curve.

ParamCurveExtrema
A parametrized curve that reports its extrema.

ParamCurveNearest
A parametrized curve that reports the nearest point.