Struct kurbo::simplify::SimplifyBezPath

pub struct SimplifyBezPath(/* private fields */);

A Bézier path which has been prepared for simplification.

See the module-level documentation for a bit more discussion of the approach, and how this struct is to be used.

Implementations

impl SimplifyBezPath

pub fn new(path: impl IntoIterator<Item = PathEl>) -> Self

Set up a new Bézier path for simplification.

Currently this is not dealing with discontinuities at all, but it could be extended to do so.