Struct kurbo::CubicBez

pub struct CubicBez {
    pub p0: Point,
    pub p1: Point,
    pub p2: Point,
    pub p3: Point,
}

A single cubic Bézier segment.

Fields
p0: Pointp1: Pointp2: Pointp3: Point
Methods

impl CubicBez

pub fn new<P: Into<Point>>(p0: P, p1: P, p2: P, p3: P) -> CubicBez

Create a new cubic Bézier segment.

pub fn to_quads(
&self,
accuracy: f64
) -> impl Iterator<Item = (f64, f64, QuadBez)>

Convert to quadratic Béziers.

The iterator returns the start and end parameter in the cubic of each quadratic segment, along with the quadratic.

Note that the resulting quadratic Béziers are not in general G1 continuous; they are optimized for minimizing distance error.