kurbo

Struct StrokeCtx

pub struct StrokeCtx {
    output: BezPath,
    forward_path: BezPath,
    backward_path: BezPath,
    result_path: BezPath,
    start_pt: Point,
    start_norm: Vec2,
    start_tan: Vec2,
    last_pt: Point,
    last_tan: Vec2,
    join_thresh: f64,
}

A structure that is used for creating strokes.

See also stroke_with.

Fields
output: BezPathforward_path: BezPathbackward_path: BezPathresult_path: BezPathstart_pt: Pointstart_norm: Vec2start_tan: Vec2last_pt: Pointlast_tan: Vec2join_thresh: f64
Implementations

impl StrokeCtx

pub fn output(&self) -> &BezPath

Return the path that defines the expanded stroke.

impl StrokeCtx

fn reset(&mut self)

impl StrokeCtx

fn finish(&mut self, style: &Stroke)

Append forward and backward paths to output.

fn finish_closed(&mut self, style: &Stroke)

Finish a closed path

fn do_join(&mut self, style: &Stroke, tan0: Vec2)

fn do_line(&mut self, style: &Stroke, tangent: Vec2, p1: Point)

fn do_cubic(&mut self, style: &Stroke, c: CubicBez, tolerance: f64)

fn do_linear( &mut self, style: &Stroke, c: CubicBez, p: [f64; 4], ref_pt: Point, ref_vec: Vec2, )

Do a cubic which is actually linear.

The p argument is the control points projected to the reference chord. The ref arguments are the inverse map of a projection back to the client coordinate space.