Enum kurbo::PathSeg

pub enum PathSeg {
    Line(Line),
    Quad(QuadBez),
    Cubic(CubicBez),
}

A segment of a Bézier path.

Variants

Line(Line)

A line segment.

Quad(QuadBez)

A quadratic bezier segment.

Cubic(CubicBez)

A cubic bezier segment.

Implementations

impl PathSeg

pub fn as_path_el(&self) -> PathEl

Get the PathEl that is equivalent to discarding the segment start point.

pub fn reverse(&self) -> PathSeg

Returns a new PathSeg describing the same path as self, but with the points reversed.

pub fn to_cubic(&self) -> CubicBez

Convert this segment to a cubic bezier.

pub fn intersect_line(&self, line: Line) -> ArrayVec<LineIntersection, 3>

Compute intersections against a line.

Returns a vector of the intersections. For each intersection, the t value of the segment and line are given.

Note: This test is designed to be inclusive of points near the endpoints of the segment. This is so that testing a line against multiple contiguous segments of a path will be guaranteed to catch at least one of them. In such cases, use higher level logic to coalesce the hits (the t value may be slightly outside the range of 0..1).

Examples

let seg = PathSeg::Line(Line::new((0.0, 0.0), (2.0, 0.0)));
let line = Line::new((1.0, 2.0), (1.0, -2.0));
let intersection = seg.intersect_line(line);
assert_eq!(intersection.len(), 1);
let intersection = intersection[0];
assert_eq!(intersection.segment_t, 0.5);
assert_eq!(intersection.line_t, 0.5);

let point = seg.eval(intersection.segment_t);
assert_eq!(point, Point::new(1.0, 0.0));

pub fn is_finite(&self) -> bool

Is this Bezier path finite?

pub fn is_nan(&self) -> bool

Is this Bezier path NaN?

pub fn min_dist(&self, other: PathSeg, accuracy: f64) -> MinDistance

Minimum distance between two PathSegs

Returns a tuple of the distance, the path time t1 of the closest point on the first PathSeg, and the path time t2 of the closest point on the second PathSeg.