kurbo

Module stroke

Structs

DashIterator 🔒

An implementation of dashing as an iterator-to-iterator transformation.

Stroke

The visual style of a stroke.

StrokeCtx

A structure that is used for creating strokes.

StrokeOpts

Options for path stroking.

Enums

Cap

Defines the shape to be drawn at the ends of a stroke.

DashState 🔒

Join

Defines the connection between two segments of a stroke.

StrokeOptLevel

Optimization level for computing stroke outlines.

Constants

DASH_ACCURACY 🔒

Functions

dash

Create a new dashing iterator.

extend_reversed 🔒

round_cap 🔒

round_join 🔒

round_join_rev 🔒

seg_to_el 🔒

square_cap 🔒

stroke

Expand a stroke into a fill.

stroke_undashed 🔒

Version of stroke expansion for styles with no dashes.

stroke_with

Expand a stroke into a fill.

Type Aliases

Dashes

Collection of values representing lengths in a dash pattern.