Struct kurbo::Arc

pub struct Arc {
    pub center: Point,
    pub radii: Vec2,
    pub start_angle: f64,
    pub sweep_angle: f64,
    pub x_rotation: f64,
}

A single arc segment.

Fields
center: Point

The arc's centre point.

radii: Vec2

The arc's radii, where the vector's x-component is the radius in the positive x direction after applying x_rotation.

start_angle: f64

The start angle in radians.

sweep_angle: f64

The angle between the start and end of the arc, in radians.

x_rotation: f64

How much the arc is rotated, in radians.

Implementations

impl Arc

pub fn append_iter(&self, tolerance: f64) -> ArcAppendIter

Create an iterator generating Bezier path elements.

The generated elemets can be append to an existing bezier path.

pub fn to_cubic_beziers<P>(self, tolerance: f64, p: P) where
P: FnMut(Point, Point, Point),

Converts an Arc into a series of cubic bezier segments.

Closure will be invoked for each segment.

impl Arc

pub fn from_svg_arc(arc: &SvgArc) -> Option<Arc>

Creates an Arc from a SvgArc.

Returns None if arc is actually a straight line.