Struct kurbo::BezPath

pub struct BezPath(_);

A path that can Bézier segments up to cubic, possibly with multiple subpaths.

Methods

impl BezPath

pub fn new() -> BezPath

Create a new path.

pub fn from_vec(v: Vec<PathEl>) -> BezPath

Create a path from a vector of path elements.

BezPath also implements FromIterator<PathEl>, so it works with collect:

// a very contrived example:
use kurbo::{BezPath, PathEl};

let path = BezPath::new();
let as_vec: Vec<PathEl> = path.into_iter().collect();
let back_to_path: BezPath = as_vec.into_iter().collect();

pub fn push(&mut self, el: PathEl)

Push a generic path element onto the path.

pub fn move_to<P: Into<Point>>(&mut self, p: P)

Push a "move to" element onto the path.

pub fn line_to<P: Into<Point>>(&mut self, p: P)

Push a "line to" element onto the path.

pub fn quad_to<P: Into<Point>>(&mut self, p1: P, p2: P)

Push a "quad to" element onto the path.

pub fn curve_to<P: Into<Point>>(&mut self, p1: P, p2: P, p3: P)

Push a "curve to" element onto the path.

pub fn close_path(&mut self)

Push a "close path" element onto the path.

pub fn elements(&self) -> &[PathEl]

Get the path elements.

pub fn iter(&self) -> impl Iterator<Item = PathEl> + '_

Returns an iterator over the path's elements.

pub fn segments(&self) -> impl Iterator<Item = PathSeg> + '_

Iterate over the path segments.

pub fn get_seg(&self, ix: usize) -> Option<PathSeg>

Get the segment at the given element index.

The element index counts PathEl elements, so for example includes an initial Moveto.

pub fn is_empty(&self) -> bool

Returns true if the path contains no segments.

pub fn apply_affine(&mut self, affine: Affine)

Apply an affine transform to the path.

pub fn nearest(&self, p: Point, accuracy: f64) -> (usize, f64, f64)

Find the nearest point.

Panics if path is empty or invalid.

Note that the returned index counts segments, not elements. Thus, the initial Moveto is not counted. For a simple path consisting of a Moveto followed by Lineto/Quadto/Cubicto elements, the element index is the segment index + 1.

Returns the index of the segment, the parameter within that segment, and the square of the distance to the point.

impl BezPath

pub fn from_path_segments(segments: impl Iterator<Item = PathSeg>) -> BezPath

Create a BezPath with segments corresponding to the sequence of PathSegs

pub fn to_svg(&self) -> String

Convert the path to an SVG path string representation.

The current implementation doesn't take any special care to produce a short string (reducing precision, using relative movement).

pub fn write_to<W: Write>(&self, writer: W) -> Result<()>

Write the SVG representation of this path to the provided buffer.

pub fn from_svg(data: &str) -> Result<BezPath, SvgParseError>

Try to parse a bezier path from an SVG path element.

This is implemented on a best-effort basis, intended for cases where the user controls the source of paths, and is not intended as a replacement for a general, robust SVG parser.