Struct kurbo::RoundedRect

pub struct RoundedRect { /* private fields */ }

A rectangle with equally rounded corners.

By construction the rounded rectangle will have non-negative dimensions and radii clamped to half size of the rect.

The easiest way to create a RoundedRect is often to create a Rect, and then call to_rounded_rect.

use kurbo::{RoundedRect, RoundedRectRadii};

// Create a rounded rectangle with a single radius for all corners:
RoundedRect::new(0.0, 0.0, 10.0, 10.0, 5.0);

// Or, specify different radii for each corner, clockwise from the top-left:
RoundedRect::new(0.0, 0.0, 10.0, 10.0, (1.0, 2.0, 3.0, 4.0));

Implementations

impl RoundedRect

pub fn new( x0: f64, y0: f64, x1: f64, y1: f64, radii: impl Into<RoundedRectRadii>, ) -> RoundedRect

A new rectangle from minimum and maximum coordinates.

The result will have non-negative width, height and radii.

pub fn from_rect(rect: Rect, radii: impl Into<RoundedRectRadii>) -> RoundedRect

A new rounded rectangle from a rectangle and corner radii.

The result will have non-negative width, height and radii.

See also Rect::to_rounded_rect, which offers the same utility.

pub fn from_points( p0: impl Into<Point>, p1: impl Into<Point>, radii: impl Into<RoundedRectRadii>, ) -> RoundedRect

A new rectangle from two Points.

The result will have non-negative width, height and radius.

pub fn from_origin_size( origin: impl Into<Point>, size: impl Into<Size>, radii: impl Into<RoundedRectRadii>, ) -> RoundedRect

A new rectangle from origin and size.

The result will have non-negative width, height and radius.

pub fn width(&self) -> f64

The width of the rectangle.

pub fn height(&self) -> f64

The height of the rectangle.

pub fn radii(&self) -> RoundedRectRadii

Radii of the rounded corners.

pub fn rect(&self) -> Rect

The (non-rounded) rectangle.

pub fn origin(&self) -> Point

The origin of the rectangle.

This is the top left corner in a y-down space.

pub fn center(&self) -> Point

The center point of the rectangle.

pub fn is_finite(&self) -> bool

Is this rounded rectangle finite?

pub fn is_nan(&self) -> bool

Is this rounded rectangle NaN?