kurbo

Struct Size

pub struct Size {
    pub width: f64,
    pub height: f64,
}

A 2D size.

Fields
width: f64

The width.

height: f64

The height.

Implementations

impl Size

pub const ZERO: Size

A size with zero width or height.

pub const INFINITY: Size

A size with width and height set to f64::INFINITY.

pub const fn new(width: f64, height: f64) -> Self

Create a new Size with the provided width and height.

pub fn max_side(self) -> f64

Returns the max of width and height.

Examples

use kurbo::Size;
let size = Size::new(-10.5, 42.0);
assert_eq!(size.max_side(), 42.0);

pub fn min_side(self) -> f64

Returns the min of width and height.

Examples

use kurbo::Size;
let size = Size::new(-10.5, 42.0);
assert_eq!(size.min_side(), -10.5);

pub fn area(self) -> f64

The area covered by this size.

pub fn is_zero_area(self) -> bool

Whether this size has zero area.

pub fn is_empty(self) -> bool

๐Deprecated since 0.11.1: use is_zero_area instead

Whether this size has zero area.

Note: a size with negative area is not considered empty.

pub fn min(self, other: Size) -> Self

Returns the component-wise minimum of self and other.

Examples

use kurbo::Size;

let this = Size::new(0., 100.);
let other = Size::new(10., 10.);

assert_eq!(this.min(other), Size::new(0., 10.));

pub fn max(self, other: Size) -> Self

Returns the component-wise maximum of self and other.

Examples

use kurbo::Size;

let this = Size::new(0., 100.);
let other = Size::new(10., 10.);

assert_eq!(this.max(other), Size::new(10., 100.));

pub fn clamp(self, min: Size, max: Size) -> Self

Returns a new size bounded by min and max.

Examples

use kurbo::Size;

let this = Size::new(0., 100.);
let min = Size::new(10., 10.,);
let max = Size::new(50., 50.);
assert_eq!(this.clamp(min, max), Size::new(10., 50.))

pub const fn to_vec2(self) -> Vec2

Convert this size into a Vec2, with width mapped to x and height mapped to y.

pub fn round(self) -> Size

Returns a new Size, with width and height rounded to the nearest integer.

Examples

use kurbo::Size;
let size_pos = Size::new(3.3, 3.6).round();
assert_eq!(size_pos.width, 3.0);
assert_eq!(size_pos.height, 4.0);
let size_neg = Size::new(-3.3, -3.6).round();
assert_eq!(size_neg.width, -3.0);
assert_eq!(size_neg.height, -4.0);

pub fn ceil(self) -> Size

Returns a new Size, with width and height rounded up to the nearest integer, unless they are already an integer.

Examples

use kurbo::Size;
let size_pos = Size::new(3.3, 3.6).ceil();
assert_eq!(size_pos.width, 4.0);
assert_eq!(size_pos.height, 4.0);
let size_neg = Size::new(-3.3, -3.6).ceil();
assert_eq!(size_neg.width, -3.0);
assert_eq!(size_neg.height, -3.0);

pub fn floor(self) -> Size

Returns a new Size, with width and height rounded down to the nearest integer, unless they are already an integer.

Examples

use kurbo::Size;
let size_pos = Size::new(3.3, 3.6).floor();
assert_eq!(size_pos.width, 3.0);
assert_eq!(size_pos.height, 3.0);
let size_neg = Size::new(-3.3, -3.6).floor();
assert_eq!(size_neg.width, -4.0);
assert_eq!(size_neg.height, -4.0);

pub fn expand(self) -> Size

Returns a new Size, with width and height rounded away from zero to the nearest integer, unless they are already an integer.

Examples

use kurbo::Size;
let size_pos = Size::new(3.3, 3.6).expand();
assert_eq!(size_pos.width, 4.0);
assert_eq!(size_pos.height, 4.0);
let size_neg = Size::new(-3.3, -3.6).expand();
assert_eq!(size_neg.width, -4.0);
assert_eq!(size_neg.height, -4.0);

pub fn trunc(self) -> Size

Returns a new Size, with width and height rounded towards zero to the nearest integer, unless they are already an integer.

Examples

use kurbo::Size;
let size_pos = Size::new(3.3, 3.6).trunc();
assert_eq!(size_pos.width, 3.0);
assert_eq!(size_pos.height, 3.0);
let size_neg = Size::new(-3.3, -3.6).trunc();
assert_eq!(size_neg.width, -3.0);
assert_eq!(size_neg.height, -3.0);

pub fn aspect_ratio(self) -> f64

Returns the aspect ratio of a rectangle with the given size.

If the width is 0, the output will be sign(self.height) * infinity. If The width and height are 0, then the output will be NaN.

pub const fn to_rect(self) -> Rect

Convert this Size into a Rect with origin (0.0, 0.0).

pub fn to_rounded_rect(self, radii: impl Into<RoundedRectRadii>) -> RoundedRect

Convert this Size into a RoundedRect with origin (0.0, 0.0) and the provided corner radius.

pub fn is_finite(self) -> bool

Is this size finite?

pub fn is_nan(self) -> bool

Is this size NaN?