Struct kurbo::TranslateScale

pub struct TranslateScale { /* private fields */ }

A transformation including scaling and translation.

If the translation is (x, y) and the scale is s, then this transformation represents this augmented matrix:

| s 0 x |
| 0 s y |
| 0 0 1 |

See Affine for more details about the equivalence with augmented matrices.

Various multiplication ops are defined, and these are all defined to be consistent with matrix multiplication. Therefore, TranslateScale * Point is defined but not the other way around.

Also note that multiplication is not commutative. Thus, TranslateScale::scale(2.0) * TranslateScale::translate(Vec2::new(1.0, 0.0)) has a translation of (2, 0), while TranslateScale::translate(Vec2::new(1.0, 0.0)) * TranslateScale::scale(2.0) has a translation of (1, 0). (Both have a scale of 2; also note that the first case can be written 2.0 * TranslateScale::translate(Vec2::new(1.0, 0.0)) as this case has an implicit conversion).

This transformation is less powerful than Affine, but can be applied to more primitives, especially including Rect.

Implementations

impl TranslateScale

pub const fn new(translation: Vec2, scale: f64) -> TranslateScale

Create a new transformation from translation and scale.

pub const fn scale(s: f64) -> TranslateScale

Create a new transformation with scale only.

pub const fn translate(t: Vec2) -> TranslateScale

Create a new transformation with translation only.

pub fn as_tuple(self) -> (Vec2, f64)

Decompose transformation into translation and scale.

pub fn inverse(self) -> TranslateScale

Compute the inverse transform.

Multiplying a transform with its inverse (either on the left or right) results in the identity transform (modulo floating point rounding errors).

Produces NaN values when scale is zero.

pub fn is_finite(&self) -> bool

Is this translate/scale finite?

pub fn is_nan(&self) -> bool

Is this translate/scale NaN?