Struct kurbo::Vec2

pub struct Vec2 {
    pub x: f64,
    pub y: f64,
}

A 2D vector.

This is intended primarily for a vector in the mathematical sense, but it can be interpreted as a translation, and converted to and from a point (vector relative to the origin) and size.

Fields
x: f64

The x-coordinate.

y: f64

The y-coordinate.

Methods

impl Vec2

pub const ZERO: Vec2

The vector (0, 0).

pub const fn new(x: f64, y: f64) -> Vec2

Create a new vector.

pub const fn to_point(self) -> Point

Convert this vector into a Point.

pub const fn to_size(self) -> Size

Convert this vector into a Size.

pub fn dot(&self, other: Vec2) -> f64

Dot product of two vectors.

pub fn cross(&self, other: Vec2) -> f64

Cross product of two vectors.

This is signed so that (0, 1) × (1, 0) = 1.

pub fn hypot(&self) -> f64

Magnitude of vector.

pub fn hypot2(&self) -> f64

Magnitude squared of vector.

pub fn atan2(&self) -> f64

Angle of vector.

If the vector is interpreted as a complex number, this is the argument. The angle is expressed in radians.

pub fn from_angle(th: f64) -> Vec2

A unit vector of the given angle.

With th at zero, the result is the positive X unit vector, and at π/2, it is the positive Y unit vector. The angle is expressed in radians.

Thus, in a Y-down coordinate system (as is common for graphics), it is a clockwise rotation, and in Y-up (traditional for math), it is anti-clockwise. This convention is consistent with Affine::rotate.

pub fn lerp(&self, other: Vec2, t: f64) -> Vec2

Linearly interpolate between two vectors.

pub fn normalize(self) -> Vec2

Returns a vector of magnitude 1.0 with the same angle as self; i.e. a unit/direction vector.

This produces NaN values when the magnitutde is 0.

pub fn round(self) -> Vec2

Returns a new Vec2 with each of x and y rounded to the nearest integer.

pub fn ceil(self) -> Vec2

Returns a new Vec2 where each of x and y, with a non-zero fractional part is rounded up to the nearest integer.

Examples

use kurbo::Vec2;

let v = Vec2::new(5.0, -1.1);
let ceil_v = v.ceil();
assert_eq!((ceil_v.x, ceil_v.y), (5.0, -1.0));

pub fn floor(self) -> Vec2

Returns a new Vec2 where each of x and y, with a non-zero fractional part is rounded down to the nearest integer.

Examples

use kurbo::Vec2;

let v = Vec2::new(4.9, -1.1);
let floor_v = v.floor();
assert_eq!((floor_v.x, floor_v.y), (4.0, -2.0));