kurbo

Trait ParamCurveNearest

pub trait ParamCurveNearest {
    // Required method
    fn nearest(&self, p: Point, accuracy: f64) -> Nearest;
}

A parameterized curve that reports the nearest point.

Required Methods

fn nearest(&self, p: Point, accuracy: f64) -> Nearest

Find the position on the curve that is nearest to the given point.

This returns a Nearest struct that contains information about the position.